third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-pause/function-name-in-callstack-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that callFrames on pause contains function name taking into account displayName and Function.name.

Set timer for test function.
Script execution paused.
callFrames.length = 4
functionName: foo.displayName
functionName: bar.displayName
functionName: baz.function.name
functionName: testFunction
Script execution resumed.
